reafator: separate post module
- separate post.rs
