+++ src/handlers/post/mod.rs
@@ -1,0 +1,8 @@
+pub mod comment;
+pub mod draft;
+pub mod post;
+pub mod share;
+pub mod stock;
+pub mod tag;
+
+pub const PAGINATES_PER: i32 = 10;
