+++ src/handlers/post/post.rs
@@ -1,4 +1,3 @@
-use persistent;
 use iron::prelude::*;
 use iron::status;
 use router::Router;
@@ -8,14 +7,15 @@ use hbs::handlebars::to_json;
 use iron::Url;
 use iron::prelude::IronResult;
 use diff;
-
 use db;
+use persistent;
+
 use env::CONFIG;
 use helper;
 use models;
 use handlers;
 
-const PAGINATES_PER: i32 = 10;
+pub const PAGINATES_PER: i32 = 10;
 // const POST_KIND: &str = "post";
 
 pub fn new_handler(req: &mut Request) -> IronResult<Response> {
@@ -559,467 +559,4 @@ pub fn update_handler(req: &mut Request) -> IronResult<Response> {
             return Ok(Response::with((status::InternalServerError)));
         }
     }
-}
-
-pub fn comment_handler(req: &mut Request) -> IronResult<Response> {
-    let conn = get_pg_connection!(req);
-    let mut login_user: models::user::User = models::user::User{..Default::default()};
-    match handlers::account::current_user(req, &conn) {
-        Ok(user) => { login_user = user; }
-        Err(e) => { error!("Errored: {:?}", e); }
-    }
-    let login_id = login_user.id;
-    if login_id == 0 {
-        return Ok(Response::with((status::Found, Redirect(helper::redirect_url("/signin")))));
-    }
-
-    let id: i32;
-    let body: String;
-
-    use params::{Params, Value};
-    {
-        let map = req.get_ref::<Params>().unwrap();
-
-        match map.find(&["id"]) {
-            Some(&Value::String(ref name)) => {
-                id = name.parse::<i32>().unwrap();
-            }
-            _ => return Ok(Response::with((status::BadRequest))),
-        }
-
-        match map.find(&["body"]) {
-            Some(&Value::String(ref name)) => {
-                body = name.to_string();
-            }
-            _ => return Ok(Response::with((status::BadRequest))),
-        }
-    }
-
-    let ref kind = req.extensions
-        .get::<Router>()
-        .unwrap()
-        .find("kind")
-        .unwrap_or("/");
-
-    let mut mentions = Vec::new();
-
-    match models::post::get_by_id(&conn, &id) {
-        Ok(post_obj) => {
-            mentions.push(post_obj.user.username);
-        }
-        Err(e) => {
-            error!("Errored: {:?}", e);
-            return Ok(Response::with((status::InternalServerError)));
-        }
-    }
-
-    match models::post::get_comments_by_post_id(&conn, &id) {
-        Ok(comments_obj) => {
-            for comment in comments_obj {
-                let username: String = comment.user.username;
-                if !mentions.contains(&username) {
-                    mentions.push(username)
-                }
-            }
-        }
-        Err(e) => {
-            error!("Errored: {:?}", e);
-            return Ok(Response::with((status::InternalServerError)));
-        }
-    }
-
-
-    match models::post::add_comment(&conn, &login_id, &id, &body) {
-        Ok(_) => {
-            let title = String::from("New comment");
-            helper::post_to_slack(&conn, &login_id, &title, &body, &id, mentions);
-            let url = Url::parse(&format!("{}/{}/show/{}", &CONFIG.team_domain, kind, id)
-                                     .to_string()).unwrap();
-            return Ok(Response::with((status::Found, Redirect(url))));
-        }
-        Err(e) => {
-            error!("Errored: {:?}", e);
-            return Ok(Response::with((status::InternalServerError)));
-        }
-    }
-}
-
-pub fn comment_update_handler(req: &mut Request) -> IronResult<Response> {
-    let conn = get_pg_connection!(req);
-    let mut login_user: models::user::User = models::user::User{..Default::default()};
-    match handlers::account::current_user(req, &conn) {
-        Ok(user) => { login_user = user; }
-        Err(e) => { error!("Errored: {:?}", e); }
-    }
-    let login_id = login_user.id;
-    if login_id == 0 {
-        return Ok(Response::with((status::Found, Redirect(helper::redirect_url("/signin")))));
-    }
-
-    let id: i32;
-    let action: String;
-    let body: String;
-    let comment: models::post::Comment;
-
-    {
-        let ref id_str = req.extensions
-            .get::<Router>()
-            .unwrap()
-            .find("id")
-            .unwrap_or("/");
-        id = id_str.parse::<i32>().unwrap();
-    }
-
-    use params::{Params, Value};
-    {
-        let map = req.get_ref::<Params>().unwrap();
-
-        match map.find(&["action"]) {
-            Some(&Value::String(ref name)) => {
-                action = name.to_string();
-            }
-            _ => return Ok(Response::with((status::BadRequest))),
-        }
-
-        match map.find(&["body"]) {
-            Some(&Value::String(ref name)) => {
-                body = name.to_string();
-            }
-            _ => return Ok(Response::with((status::BadRequest))),
-        }
-    }
-
-    let ref kind = req.extensions
-        .get::<Router>()
-        .unwrap()
-        .find("kind")
-        .unwrap_or("/");
-
-    match models::post::get_comment_by_id(&conn, &id) {
-        Ok(db_comment) => {
-            comment = db_comment;
-            if comment.user_id != login_id {
-                return Ok(Response::with((status::Forbidden)));
-            }
-        }
-        Err(e) => {
-            error!("Errored: {:?}", e);
-            return Ok(Response::with((status::InternalServerError)));
-        }
-    }
-
-    if action == "update" {
-        match models::post::update_comment_by_id(&conn, &id, &body) {
-            Ok(_) => {
-                let url = Url::parse(&format!("{}/{}/show/{}", &CONFIG.team_domain, kind, comment.post_id)
-                                         .to_string()).unwrap();
-                return Ok(Response::with((status::Found, Redirect(url))));
-            }
-            Err(e) => {
-                error!("Errored: {:?}", e);
-                return Ok(Response::with((status::InternalServerError)));
-            }
-        }
-    }
-    if action == "delete" {
-        match models::post::delete_comment_by_id(&conn, &id) {
-            Ok(_) => {
-                let url = Url::parse(&format!("{}/{}/show/{}", &CONFIG.team_domain, kind, comment.post_id)
-                                         .to_string()).unwrap();
-                return Ok(Response::with((status::Found, Redirect(url))));
-            }
-            Err(e) => {
-                error!("Errored: {:?}", e);
-                return Ok(Response::with((status::InternalServerError)));
-            }
-        }
-    }
-    return Ok(Response::with((status::InternalServerError)));
-}
-
-pub fn stock_handler(req: &mut Request) -> IronResult<Response> {
-    let conn = get_pg_connection!(req);
-    let mut login_user: models::user::User = models::user::User{..Default::default()};
-    match handlers::account::current_user(req, &conn) {
-        Ok(user) => { login_user = user; }
-        Err(e) => { error!("Errored: {:?}", e); }
-    }
-    let login_id = login_user.id;
-    if login_id == 0 {
-        return Ok(Response::with((status::Found, Redirect(helper::redirect_url("/signin")))));
-    }
-
-    let ref kind = req.extensions
-        .get::<Router>()
-        .unwrap()
-        .find("kind")
-        .unwrap_or("/");
-
-    let ref id_str = req.extensions
-        .get::<Router>()
-        .unwrap()
-        .find("id")
-        .unwrap_or("/");
-    let id = id_str.parse::<i32>().unwrap();
-
-    match models::post::stock_post(&conn, &login_id, &id) {
-        Ok(_) => {
-            let url = Url::parse(&format!("{}/{}/show/{}", &CONFIG.team_domain, kind, id)
-                    .to_string())
-                    .unwrap();
-            return Ok(Response::with((status::Found, Redirect(url))));
-        }
-        Err(e) => {
-            error!("Errored: {:?}", e);
-            return Ok(Response::with((status::InternalServerError)));
-        }
-    }
-}
-
-pub fn unstock_handler(req: &mut Request) -> IronResult<Response> {
-    let conn = get_pg_connection!(req);
-    let mut login_user: models::user::User = models::user::User{..Default::default()};
-    match handlers::account::current_user(req, &conn) {
-        Ok(user) => { login_user = user; }
-        Err(e) => { error!("Errored: {:?}", e); }
-    }
-    let login_id = login_user.id;
-    if login_id == 0 {
-        return Ok(Response::with((status::Found, Redirect(helper::redirect_url("/signin")))));
-    }
-
-    let ref kind = req.extensions
-        .get::<Router>()
-        .unwrap()
-        .find("kind")
-        .unwrap_or("/");
-
-    let ref id_str = req.extensions
-        .get::<Router>()
-        .unwrap()
-        .find("id")
-        .unwrap_or("/");
-    let id = id_str.parse::<i32>().unwrap();
-
-    match models::post::stock_remove(&conn, &login_id, &id) {
-        Ok(_) => {
-            let url = Url::parse(&format!("{}/{}/show/{}", &CONFIG.team_domain, kind, id)
-                    .to_string())
-                    .unwrap();
-            return Ok(Response::with((status::Found, Redirect(url))));
-        }
-        Err(e) => {
-            error!("Errored: {:?}", e);
-            return Ok(Response::with((status::InternalServerError)));
-        }
-    }
-}
-
-pub fn share_handler(req: &mut Request) -> IronResult<Response> {
-    let conn = get_pg_connection!(req);
-    let mut login_user: models::user::User = models::user::User{..Default::default()};
-    match handlers::account::current_user(req, &conn) {
-        Ok(user) => { login_user = user; }
-        Err(e) => { error!("Errored: {:?}", e); }
-    }
-    let login_id = login_user.id;
-    if login_id == 0 {
-        return Ok(Response::with((status::Found, Redirect(helper::redirect_url("/signin")))));
-    }
-
-    let ref kind = req.extensions
-        .get::<Router>()
-        .unwrap()
-        .find("kind")
-        .unwrap_or("/");
-
-    let ref id_str = req.extensions
-        .get::<Router>()
-        .unwrap()
-        .find("id")
-        .unwrap_or("/");
-    let id = id_str.parse::<i32>().unwrap();
-
-    match models::post::share_post(&conn, &id) {
-        Ok(_) => {
-            let url = Url::parse(&format!("{}/{}/show/{}", &CONFIG.team_domain, kind, id)
-                    .to_string())
-                    .unwrap();
-            return Ok(Response::with((status::Found, Redirect(url))));
-        }
-        Err(e) => {
-            error!("Errored: {:?}", e);
-            return Ok(Response::with((status::InternalServerError)));
-        }
-    }
-}
-
-pub fn stocked_list_handler(req: &mut Request) -> IronResult<Response> {
-    let conn = get_pg_connection!(req);
-    let mut login_user: models::user::User = models::user::User{..Default::default()};
-    match handlers::account::current_user(req, &conn) {
-        Ok(user) => { login_user = user; }
-        Err(e) => { error!("Errored: {:?}", e); }
-    }
-    let login_id = login_user.id;
-    if login_id == 0 {
-        return Ok(Response::with((status::Found, Redirect(helper::redirect_url("/signin")))));
-    }
-
-    let page_param: String;
-
-    {
-        use params::{Params, Value};
-        let map = req.get_ref::<Params>().unwrap();
-        match map.get("page") {
-            Some(&Value::String(ref name)) => {
-                page_param = name.to_string();
-            }
-            _ => page_param = "1".to_string(),
-        }
-    }
-
-    let mut resp = Response::new();
-
-    #[derive(Serialize, Debug)]
-    struct Data {
-        logged_in: bool,
-        login_user: models::user::User,
-        posts: Vec<models::post::Post>,
-        current_page: i32,
-        total_page: i32,
-        next_page: i32,
-        prev_page: i32,
-    }
-
-    let mut page = page_param.parse::<i32>().unwrap();
-    let offset = (page - 1) * PAGINATES_PER;
-    let limit = PAGINATES_PER;
-
-    let posts: Vec<models::post::Post>;
-    let count: i32;
-
-    match models::post::stocked_list(&conn, &login_id, &offset, &limit) {
-        Ok(posts_db) => {
-            posts = posts_db;
-        }
-        Err(e) => {
-            error!("Errored: {:?}", e);
-            return Ok(Response::with((status::InternalServerError)));
-        }
-    }
-
-    match models::post::stocked_count(&conn, &login_id) {
-        Ok(count_db) => {
-            count = count_db;
-        }
-        Err(e) => {
-            error!("Errored: {:?}", e);
-            return Ok(Response::with((status::InternalServerError)));
-        }
-    }
-
-    if page == 0 {
-        page = 1;
-    }
-    let data = Data {
-        logged_in: login_id != 0,
-        login_user: login_user,
-        posts: posts,
-        current_page: page,
-        total_page: count / PAGINATES_PER + 1,
-        next_page: page + 1,
-        prev_page: page - 1,
-    };
-
-    resp.set_mut(Template::new("stock/list", to_json(&data)))
-        .set_mut(status::Ok);
-    return Ok(resp);
-}
-
-pub fn draft_list_handler(req: &mut Request) -> IronResult<Response> {
-    let conn = get_pg_connection!(req);
-    let mut login_user: models::user::User = models::user::User{..Default::default()};
-    match handlers::account::current_user(req, &conn) {
-        Ok(user) => { login_user = user; }
-        Err(e) => { error!("Errored: {:?}", e); }
-    }
-    let login_id = login_user.id;
-    if login_id == 0 {
-        return Ok(Response::with((status::Found, Redirect(helper::redirect_url("/signin")))));
-    }
-
-    let mut resp = Response::new();
-
-    #[derive(Serialize, Debug)]
-    struct Data {
-        logged_in: bool,
-        login_user: models::user::User,
-        posts: Vec<models::post::Post>,
-    }
-
-    let posts: Vec<models::post::Post>;
-
-    match models::post::draft_list(&conn, &login_id) {
-        Ok(posts_db) => {
-            posts = posts_db;
-        }
-        Err(e) => {
-            error!("Errored: {:?}", e);
-            return Ok(Response::with((status::InternalServerError)));
-        }
-    }
-    let data = Data {
-        logged_in: login_id != 0,
-        login_user: login_user,
-        posts: posts,
-    };
-
-    resp.set_mut(Template::new("draft/list", to_json(&data)))
-        .set_mut(status::Ok);
-    return Ok(resp);
-
-}
-
-pub fn tag_list_handler(req: &mut Request) -> IronResult<Response> {
-    let conn = get_pg_connection!(req);
-    let mut login_user: models::user::User = models::user::User{..Default::default()};
-    match handlers::account::current_user(req, &conn) {
-        Ok(user) => { login_user = user; }
-        Err(e) => { error!("Errored: {:?}", e); }
-    }
-    let login_id = login_user.id;
-    if login_id == 0 {
-        return Ok(Response::with((status::Found, Redirect(helper::redirect_url("/signin")))));
-    }
-
-    let mut resp = Response::new();
-
-    #[derive(Serialize, Debug)]
-    struct Data {
-        logged_in: bool,
-        login_user: models::user::User,
-        tags: Vec<models::tag::Tag>,
-    }
-
-    let tags: Vec<models::tag::Tag>;
-
-    match models::tag::tag_list(&conn) {
-        Ok(tags_db) => {
-            tags = tags_db;
-        }
-        Err(e) => {
-            error!("Errored: {:?}", e);
-            return Ok(Response::with((status::InternalServerError)));
-        }
-    }
-    let data = Data {
-        logged_in: login_id != 0,
-        login_user: login_user,
-        tags: tags,
-    };
-
-    resp.set_mut(Template::new("tag/list", to_json(&data)))
-        .set_mut(status::Ok);
-    return Ok(resp);
 }
